# Inline bits of pdf crate for better performance