--- a/crates/cracker/Cargo.toml
+++ b/crates/cracker/Cargo.toml
@@ -9,3 +9,4 @@ authors.workspace = true
 [dependencies]
 log.workspace = true
 pdf.workspace = true
+anyhow.workspace = true